replace `util` helper with updated variant for rebased
