+++ bindings/wasm/identity_wasm/examples/src/1_advanced/7_status_list_2021.ts
@@ -21,7 +21,7 @@ import {
     getClientAndCreateAccount,
     getMemstorage,
     NETWORK_URL,
-} from '../utils_alpha';
+} from '../util';
 
 export async function statusList2021() {
     // ===========================================================================
